Fix off-by-one bug causing RichText to copy 1 char more than selected
This annoying bug has historically been most noticeable with the console window (`CConsolePanel`), which uses a `RichText` to display the history buffer. I can't count the number of times I've seen people post console excerpts (e.g. to a forum post when reporting a bug, or just to chat during a game) that are supposed to include just the line(s) the user selected, but which end up additionally including the first character from the line after that, which was *not* part of the actual selected text.

`RichText::GetSelectedRange` returns (by-ref) the values stored in `_select[0]` and `_select[1]` as `cx0` and `cx1` (sorted such that `cx0` is the selection start index and `cx1` is the selection end index). The important thing to note here is that the range of selected text is EXCLUSIVE of `cx1`; that is, the selected index range is `[cx0, cx1)`; or `[cx0, cx1-1]` if you prefer. This is most obviously demonstrated by a quick glance at `RichText::SelectAllText`, which sets the selection end position to `m_TextStream.Count()`. So when the for-loop in `RichText::CopySelected` iterates from `x0` to `x1` INCLUSIVE, it erroneously includes an extra character at the end which the user did not actually select.

This patch changes the for-loop termination condition in `RichText::CopySelected` so that the last character appended to the vector is the character at index `x1-1`, rather than the one at index `x1`.